test: move sync tests to `astro sync`, add file write test
