+++ packages/astro/test/content-collections.test.js
@@ -1,4 +1,3 @@
-import * as fs from 'node:fs';
 import * as devalue from 'devalue';
 import * as cheerio from 'cheerio';
 import { expect } from 'chai';
@@ -6,55 +5,6 @@ import { loadFixture } from './test-utils.js';
 import testAdapter from './test-adapter.js';
 
 describe('Content Collections', () => {
-	describe('Type generation', () => {
-		let fixture;
-		before(async () => {
-			fixture = await loadFixture({ root: './fixtures/content-collections/' });
-		});
-
-		it('Writes types to `.astro`', async () => {
-			let writtenFiles = {};
-			const fsMock = {
-				...fs,
-				promises: {
-					...fs.promises,
-					async writeFile(path, contents) {
-						writtenFiles[path] = contents;
-					},
-				},
-			};
-			await fixture.sync({ fs: fsMock });
-
-			const expectedTypesFile = new URL('.astro/types.d.ts', fixture.config.root).href;
-			expect(writtenFiles).to.haveOwnProperty(expectedTypesFile);
-			// smoke test `astro check` asserts whether content types pass.
-			expect(writtenFiles[expectedTypesFile]).to.include(
-				`declare module 'astro:content' {`,
-				'Types file does not include `astro:content` module declaration'
-			);
-		});
-
-		it('Adds type reference to `src/env.d.ts`', async () => {
-			let writtenFiles = {};
-			const typesEnvPath = new URL('env.d.ts', fixture.config.srcDir).href;
-			const fsMock = {
-				...fs,
-				promises: {
-					...fs.promises,
-					async writeFile(path, contents) {
-						writtenFiles[path] = contents;
-					},
-				},
-			};
-			await fixture.sync({ fs: fsMock });
-
-			expect(writtenFiles, 'Did not try to update env.d.ts file.').to.haveOwnProperty(typesEnvPath);
-			expect(writtenFiles[typesEnvPath]).to.include(
-				`/// <reference path="../.astro/types.d.ts" />`
-			);
-		});
-	});
-
 	describe('Query', () => {
 		let fixture;
 		before(async () => {
